ash: HACK: "local" at global scope is nonfatal.
Incorrect syntax in legacy software left a "local" keyword at global
scope in boot-time scripting. The product in question needs to be able
to downgrade to older versions, and so this "local" keyword must not
cause a syntax error.

This commit modifies or reverts ef2386b80a and b8ab27bf53.

Signed-off-by: Rick Ramstetter <rick@anteaterllc.com>